Correct docblock types and annotate view templates
Fixes the type declarations PHPStan could resolve without touching behaviour:
return and property types that named the wrong type, two class names missing
their leading backslash, and explicit `return null;` where a docblock declares
a return type. The view templates gain a header docblock naming the variables
the including method supplies, since PHPStan analyses them standalone.

Two changes reach runtime and are no-ops in practice: Helper::handle_currency()
now casts its result to int, which the SDK's `?int` setters already coerced, and
a WC_Order_Refund::set_amount() call passes '0' rather than 0.

Integration suite: 104 tests, 152 assertions, green.

+++ src/Helper.php
@@ -55,10 +55,10 @@ class Helper {
 	 * Currency specific formattings
 	 *
 	 * @param int|double $sum The sum to format.
-	 * @return integer
+	 * @return int
 	 */
 	public function handle_currency( $sum ) {
-		return round( $sum * 100 );
+		return (int) round( $sum * 100 );
 	}
 
 	/**
